+++ tests/unit/test_utils.py
@@ -80,10 +80,6 @@ class UtilsTests(SparkBaseTest):
     test_csv_path = "tests/test_data/example_csv.csv"
     test_csv_custom_delim_path = "tests/test_data/example_csv_custom_delimiter.csv"
     TEST_ASCWDS_WORKPLACE_FILE = "tests/test_data/tmp-workplace"
-    example_csv_for_schema_tests = "tests/test_data/example_csv_for_schema_tests.csv"
-    example_csv_for_schema_tests_extra_column = (
-        "tests/test_data/example_csv_for_schema_tests_extra_column.csv"
-    )
     example_csv_for_schema_tests_with_datetype = (
         "tests/test_data/example_csv_for_schema_tests_with_datetype.csv"
     )
@@ -91,9 +87,6 @@ class UtilsTests(SparkBaseTest):
 
     def setUp(self):
         self.df = self.spark.read.csv(self.test_csv_path, header=True)
-        self.df_with_extra_col = self.spark.read.csv(
-            self.example_csv_for_schema_tests_extra_column, header=True
-        )
         self.pir_cleaned_test_df: DataFrame = self.spark.createDataFrame(
             data=UtilsData.cqc_pir_rows,
             schema=UtilsSchema.cqc_pir_schema,
@@ -113,102 +106,6 @@ class GeneralUtilsTests(UtilsTests):
         df = utils.read_csv(self.test_csv_path)
         self.assertEqual(df.columns, ["col_a", "col_b", "col_c", "date_col"])
         self.assertEqual(df.count(), 3)
-
-    def test_read_csv_with_defined_schema(self):
-        schema = StructType(
-            [
-                StructField("string_field", StringType(), True),
-                StructField("integer_field", IntegerType(), True),
-                StructField("float_field", FloatType(), True),
-            ]
-        )
-
-        df = utils.read_csv_with_defined_schema(
-            self.example_csv_for_schema_tests, schema
-        )
-        self.assertEqual(df.columns[0], "string_field")
-        self.assertEqual(df.columns[1], "integer_field")
-        self.assertEqual(df.columns[2], "float_field")
-        row_one = df.collect()[0]
-        assert isinstance(row_one.string_field, str)
-        assert isinstance(row_one.integer_field, int)
-        assert isinstance(row_one.float_field, float)
-
-    def test_read_csv_with_defined_schema_with_null_values_in_csv(self):
-        schema = StructType(
-            [
-                StructField("string_field", StringType(), False),
-                StructField("integer_field", IntegerType(), False),
-                StructField("float_field", FloatType(), False),
-            ]
-        )
-
-        df = utils.read_csv_with_defined_schema(
-            self.example_csv_for_schema_tests, schema
-        )
-        self.assertEqual(df.columns[0], "string_field")
-        self.assertEqual(df.columns[1], "integer_field")
-        self.assertEqual(df.columns[2], "float_field")
-        row_two = df.collect()[1]
-        assert isinstance(row_two.string_field, type(None))
-        assert isinstance(row_two.integer_field, type(None))
-        assert isinstance(row_two.float_field, type(None))
-
-    def test_read_csv_with_defined_schema_with_column_missing_in_csv(self):
-        schema = StructType(
-            [
-                StructField("string_field", StringType(), False),
-                StructField("integer_field", IntegerType(), False),
-                StructField("float_field", FloatType(), False),
-                StructField("missing_field", StringType(), True),
-            ]
-        )
-
-        df = utils.read_csv_with_defined_schema(
-            self.example_csv_for_schema_tests, schema
-        )
-        self.assertEqual(df.columns[0], "string_field")
-        self.assertEqual(df.columns[1], "integer_field")
-        self.assertEqual(df.columns[2], "float_field")
-        self.assertEqual(df.columns[3], "missing_field")
-
-    def test_read_csv_with_defined_schema_with_extra_column_in_csv(self):
-        schema = StructType(
-            [
-                StructField("string_field", StringType(), False),
-                StructField("integer_field", IntegerType(), False),
-                StructField("float_field", FloatType(), False),
-            ]
-        )
-
-        df_with_no_schema = self.df_with_extra_col
-
-        df = utils.read_csv_with_defined_schema(
-            self.example_csv_for_schema_tests_extra_column, schema
-        )
-        self.assertEqual(df.columns[0], "string_field")
-        self.assertEqual(df.columns[1], "integer_field")
-        self.assertEqual(df.columns[2], "float_field")
-
-        self.assertTrue(len(df.columns) < len(df_with_no_schema.columns))
-
-    def test_read_csv_with_defined_schema_where_there_is_incorrect_value_type(self):
-        schema = StructType(
-            [
-                StructField("string_field", IntegerType(), False),
-                StructField("integer_field", StringType(), False),
-                StructField("float_field", FloatType(), False),
-            ]
-        )
-
-        df = utils.read_csv_with_defined_schema(
-            self.example_csv_for_schema_tests, schema
-        )
-
-        row_one = df.collect()[0]
-        assert isinstance(row_one.string_field, type(None))
-        assert isinstance(row_one.integer_field, str)
-        assert isinstance(row_one.float_field, float)
 
     def test_read_with_custom_delimiter(self):
         df = utils.read_csv(self.test_csv_custom_delim_path, "|")